PRSD-353: Moves enums package into constants
