+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/constants/enums/RegistrationNumberType.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.enums
+package uk.gov.communities.prsdb.webapp.constants.enums
 
 enum class RegistrationNumberType {
     PROPERTY,
