+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/controllers/ExampleStartPageController.kt
@@ -5,8 +5,8 @@ import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import uk.gov.communities.prsdb.webapp.constants.SERVICE_NAME
+import uk.gov.communities.prsdb.webapp.constants.enums.RegistrationNumberType
 import uk.gov.communities.prsdb.webapp.database.entity.RegistrationNumber
-import uk.gov.communities.prsdb.webapp.enums.RegistrationNumberType
 import uk.gov.communities.prsdb.webapp.services.RegistrationNumberService
 
 @Controller
